+++ snapshot-tests-documentation/src/test/java/de/skuzzle/test/snapshots/snippets/SnapshotTestOptionsTest.java
@@ -9,6 +9,14 @@ import org.junit.jupiter.api.Test;
 @EnableSnapshotTests
 class SnapshotTestOptionsTest {
 
+    @Test
+    // tag::normalizeLineEndingsGit[]
+    @SnapshotTestOptions(normalizeLineEndings = SnapshotTestOptions.NormalizeLineEndings.GIT)
+    // end::normalizeLineEndingsGit[]
+    void normalizeLineEndingsGit(Snapshot snapshot) throws Exception {
+
+    }
+
     // tag::alwaysPersistActualResult[]
     @Test
     @SnapshotTestOptions(alwaysPersistActualResult = true)