+++ readme/RELEASE_NOTES.md
@@ -23,6 +23,7 @@
 * [#60](https://github.com/skuzzle/snapshot-tests/issues/60): Internal API for plugging in test-framework specific behavior
 * [#80](https://github.com/skuzzle/snapshot-tests/issues/80): Use spotless and reformat whole code base
 * [#88](https://github.com/skuzzle/snapshot-tests/issues/88): Extract diff creation into own module
+* [#44](https://github.com/skuzzle/snapshot-tests/issues/44)/[#90](https://github.com/skuzzle/snapshot-tests/issues/90): Add possibility to normalize line endings (according to local git config)
 
 * Update to jsonassert `1.5.1` (coming from `1.5.0`)